+++ app/src/main/java/com/example/calculator/application/expressionhistory/ExpressionHistoryActivity.kt
@@ -1,32 +1,22 @@
-package com.example.calculator.application.historyexpression
+package com.example.calculator.application.expressionhistory
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.fragment.app.Fragment
 import com.example.calculator.R
-import com.example.calculator.application.main.AboutDialog
-import kotlinx.android.synthetic.main.activity_history_expression.*
 
 class ExpressionHistoryActivity : AppCompatActivity() {
 
-    var viewModel = ExpressionHistoryViewModel()
+    var fragment = ExpressionHistoryFragment()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_history_expression)
 
-        rvExpressionHistory.layoutManager = LinearLayoutManager(this)
-
-        viewModel.observerRecords.register {
-            it.reverse()
-            var adapter = ExpressionHistoryAdapter(it, this)
-            rvExpressionHistory.adapter = adapter
-        }
-
-        viewModel.loadData()
+        fragment = supportFragmentManager.fragments[0] as ExpressionHistoryFragment
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -39,7 +29,7 @@ class ExpressionHistoryActivity : AppCompatActivity() {
 
         val id= item.itemId
         if (id == R.id.clear) {
-            viewModel.clearData()
+            fragment.clearData()
         }
 
         return super.onOptionsItemSelected(item)
