+++ app/src/main/res/layout/activity_main.xml
@@ -8,27 +8,9 @@
     android:gravity="bottom"
     tools:context=".application.main.MainActivity">
 
-    <TextView
-        android:id="@+id/etResult"
+    <fragment android:name="com.example.calculator.application.main.MainFragment"
+        android:id="@+id/main_fragment"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="3"
-        android:inputType="none"
-        android:gravity="right|bottom"
-        android:textIsSelectable="true"/>
-
-    <EditText
-        android:id="@+id/etExpression"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="2"
-        android:gravity="right|bottom"
-        android:textIsSelectable="true"/>
-
-    <GridLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/glContent" >
-    </GridLayout>
+        android:layout_height="match_parent" />
 
 </LinearLayout>
