+++ app/src/main/res/layout/fragment_expression_history.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".application.expressionhistory.ExpressionHistoryActivity">
+    tools:context=".application.expressionhistory.ExpressionHistoryFragment">
 
-    <fragment android:name="com.example.calculator.application.expressionhistory.ExpressionHistoryFragment"
-        android:id="@+id/history_fragment"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rvExpressionHistory"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
