+++ app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:gravity="bottom"
+    tools:context=".application.main.MainFragment">
+
+    <TextView
+        android:id="@+id/etResult"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="3"
+        android:inputType="none"
+        android:gravity="right|bottom"
+        android:textIsSelectable="true"/>
+
+    <EditText
+        android:id="@+id/etExpression"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="2"
+        android:gravity="right|bottom"
+        android:textIsSelectable="true"/>
+
+    <GridLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/glContent" >
+    </GridLayout>
+
+</LinearLayout>
